nav items active, remove visual guidelines main
