+++ playbook-website/app/javascript/components/Sidebar/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {Nav, NavItem, Caption} from 'playbook-ui'
+import { Nav, NavItem } from 'playbook-ui'
 
 export const MENU_ITEMS = [
   "Colors",
@@ -18,17 +18,14 @@ export const MENU_ITEMS = [
 ];
 
 const Sidebar = () => {
-  
+  const regex = /(?:(?:[^/]*\/){2})(.*)/;
   return (
     <>
-    <Caption marginY="md" 
-             marginLeft="md" >
-      <a style={{color: 'unset'}} href="https://playbook.powerapp.cloud/visual_guidelines">Visual Guidelines</a>
-    </Caption>
     <Nav variant="subtle">
         {
             MENU_ITEMS.map(item => (
                 <NavItem
+                active={item.toLowerCase().replace(/[\sA-Z-]+/g, (match) => " ".replace(/[\s-]/g, '_')) == window.location.pathname.match(regex)[1]}
                 link={`/visual_guidelines/${item.toLowerCase().replace(/[\sA-Z-]+/g, (match) => " ".replace(/[\s-]/g, '_'))}`}
                 text={item}
                 />
