+++ playbook-website/app/javascript/packs/application.js
@@ -19,7 +19,6 @@ import KitDocs from '../components/KitDocs'
 import PbKitReact from '../components/PbKitReact'
 import PbKitFetch from '../components/PbKitFetch'
 import PbKitPlayground from '../components/PbKitPlayground'
-import VisualGuidelinesIndividual from '../components/VisualGuidelinesIndividual'
 
 WebpackerReact.setup({
   DarkModeToggle,
@@ -30,7 +29,6 @@ WebpackerReact.setup({
   PbKitReact,
   PbKitFetch,
   PbKitPlayground,
-  VisualGuidelinesIndividual,
 })
 
 // Produce image assets
