+++ playbook-website/app/views/pages/visual_guidelines.html.erb
@@ -1,16 +1,4 @@
 <div class="visual_guidelines">
-  <%= pb_rails("background", props: { background_color: "dark", padding: "xl", classname: "visual_guidelines_hero" }) do %>
-    <%= pb_rails("flex", props: { classname: "header-text", orientation: "column", horizontal: "center"}) do %>
-      <div>
-        <%= pb_rails("title", props: { classname: "dark", margin_bottom: "lg", text: "Our Visual Language", tag: "h1", size: 1 }) %>
-        <%= pb_rails("body", props: {
-          classname: "dark",
-          text: "This page outlines the guiding principles that inform Playbook Design System's past, present, and future development.
-          These core concepts are the basis upon which every higher order design element is built.  Some of these principles have been abstracted into sass variables listed below."
-          }) %>
-      </div>
-    <% end %>
-  <% end %>
   <%= pb_rails("background", props: { classname: "token-wrapper", background_color: "light", padding: "xl" }) do %>
     <%= react_component("VisualGuidelines", examples: @kit_examples_json) %>
   <% end %>
