--- playbook-website/app/views/pages/visual_guidelines_individual.html.erb
@@ -1,5 +1,0 @@
-<div class="visual_guidelines">
-  <%= pb_rails("background", props: { classname: "token-wrapper", background_color: "light", padding: "xl" }) do %>
-    <%= react_component("VisualGuidelinesIndividual", examples: @kit_examples_json) %>
-  <% end %>
-</div>
